+++ NEWS.md
@@ -1,3 +1,9 @@
+# mnirs 0.1.1
+
+* Fix formatting error in README.Rmd.
+* Remove obsolete code in `resample_mnirs.R`.
+* Add some missing `test_that()` for `replace_mnirs()` and `resample_mnirs()`.
+
 # mnirs 0.1.0 initial release
 
 * Initial release of reading and cleaning functions.